Deprecates resolved_concrete_value from struct UserValueReference in mojom_types.mojom.
Also adds a comment indicating that we want to add the field resolved_concrete_value to the struct DeclaredConstant.

Also delete the function EnumValue.Int32Value() as what it was doing is incorrect: It was setting the int32 value for enumvalues that happened to have them explicitly initialized.

For now the serialized output of the parser will not have either resolved_concrete_value or EnumValue.int_value set.

R=azani@chromium.org, azani
BUG=#461

Review URL: https://codereview.chromium.org/1442593004 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: d2957747ae2657de02996f72ba211b8bb750a1f3