Update the glue between the new mojom parser and the generators to handle transitive imports.
We are running into an issue where the translation layer is incapable of handling cases where when translating module A which imports module B, which imports module C, there is a reference to a type in module C. This problem is fixed here.

In addition, the Dart and Python bindings generators modified the modules passed to them as a side-effect and depended upon those side-effects. Since those dependencies are not supported by the new binding generator, we have to pass in more information. (See Module.transitive_imports.)

R=rudominer@chromium.org, rudominer
BUG= #461

Review URL: https://codereview.chromium.org/1445293002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: c8b67932cbbc4085a7f9d7f5a30b55867c8c4c85