New Mojom Compiler:  Change the type of the |Attributes.value| field from string to LiteralValue.
BUG=#561
R=azani@chromium.org, azani

Note that because of issue #390 I needed to move the declaration of struct Attribute above the declaration of struct DeclarationData in the file mojom_types.mojom.

This patch changes the type of the |value| field in the |Attributes| struct in mojom_types.mojom from string to LiteralValue.

In the Go-based parser we were already representing attribute values as LiteralValues but we were throwing the type information
away at serialization time. This patch corrects that. (To be precise we were previously representing an attribute value as
a ConcreteValue which is more general than a LiteralValue. In this patch we have retreated to representing it as a LiteralValue.)

In the Python-based glue layer we were assuming the value was a string and setting the module object's attribute value to
be directly equal to this string. However the existing code generators are expecting an attribute value to be typed.
(The old parser simply runs the Python eval() function on the input token.) The specific issue I ran into is that the code generators
expect the min_version attribute to be an integer. This patch changes the Python glue layer to expect a mojom_types.Attribute value to
be a LiteralValue and it uses the .data field to populate the module object's attribute value. This preserves the type.

Review URL: https://codereview.chromium.org/1469213002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: e3b0acd64f56cf37eb952a61cb0d60f30008182a