Initialize local pointers to null in bindings unit tests
Some smart-but-not-too-smart compilers like gcc 4.9.3 generate
maybe-uninitialized errors for this pattern (in release mode):

void foo() {
  int* local;
  CallThingThatPopulates(&local);
}

which we trigger a lot in the bindings unit tests when calling functions
that use pointer-to-pointer as out params. The compiler is (probably)
wrong that these values could be uninitialized, but explicitly
initializing the local pointers is probably a good idea anyway to get
cleaner errors in failure cases if nothing else.

R=vardhan@google.com

Review URL: https://codereview.chromium.org/1484243002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 3df4e31674c4e9514878b5197c186dbd1661d739