Dart: Fix flaky test failures
- Adds missing .mojom.dart to test package
- Stops passing bad command line flag

The compile_all_interfaces test was flaky due to trying to pass the non-existent --compile_all command line flag. Since the flag would only make it to the VM if the test was the first to call DartController::InitVmIfNeeded() it has been failing only rarely.

I don't know the reason for the flaky failure of the bindings generation test, but based on the error message, it looks like the the .mojom in this CL is missing a DartPacakge annotation, so hopefully adding that will clear things up.

BUG=
R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/1494473002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: f64068ccd625cece496dcbc3b6457b72327a5516