Dart: Fixes nested union encoding.
Also adds default cases to union encoding and decoding.

The bug was exposed by setting the correct length for the ByteData view
for the message payload. If you do that then the union encoding gives
back a message that is too short causing pointers in the payload to
point beyond the end of the message. The bug is fixed by reserving
space for the union only after calculating where the pointer to it
should point.

BUG=
R=jimbe@chromium.org

Review URL: https://codereview.chromium.org/1510663004 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 8e14b444560d5e1c7edcad7e7ad57a328806bd1b