Revert "Rename the name of the platform to android-arm-official as in with cloud storage."
This reverts commit 1c7d44d0cb6444086fec756566b916025a13f627.

The official build should not be downloaded and used independantly. It
is there only to be pushed to the play store. If someone wants to use
the official version, they should get it from the play store.. In
particular, the official one is not a debug version of the shell. The
dev version is the one that should be downloaded and locally installed.

R=tonyg@chromium.org

Review URL: https://codereview.chromium.org/1528443002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: a5c0426375deca641eb2251800f9ef827b565ded