New Mojom compiler: Use the new compiler by default.
The new version of the Mojom compiler is now used by default.
We are not yet deleting the old version and if a user wishes he may use the old version by setting the gn variable use_old_mojom_compiler=true.
(We plan to send out a public service announcement describing this and other aspects of the new compiler prior to committing this patch.)

We also rename the flag --python-bindings-dir to
--python-sdk-dir because it is more accurate and we eliminate the redundant setting of a
default value for this. Now the default is only set in run_code_generators.py.

BUG=#593
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1551423002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 1007bff8daf2df78b959e374554ad1c64f518688