When deleting files in the mojom dart generator, avoid race conditions.
The mojom dart generator's approach to deleting files was to check if
the file exists and then try to delete it. This is problematic since the
file can be removed before the deletion operation occures. Instead, we
attempt to delete the file and ignore the error if the file did not
exist. This is also arguable more pythonic.

R=rudominer@chromium.org, rudominer
BUG=

Review URL: https://codereview.chromium.org/1564573002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 287583c86a9d251e53057755337352218106c35f