Mojom compiler: Store and serialize |specified_file_name| for top-level files.
This is part of the project to allow the code generators to optionally only generate code for the *requested* mojom files as opposed to generating
code for all mojom files in the import graph corresponding to the requested files. The front-end returns a graph containing all files and
the back-end needs to be able to determine whether or not a file in that graph was a requested (a.k.a top-level) file or a file in the graph
only due to imports.

Our current solution (from https://codereview.chromium.org/1566203002/ and https://codereview.chromium.org/1568123002/) is a hack
because it will fail of two files in the import graph have the same short name.

In this new solution we mark each file in file graph returned from the front-end with information about whether or not it is a top-level file.

* Adds a |SpecifiedFileName| field to MojomFile struct and populates it for top-level files, but not for files that are only in the file graph because of an import.

* Uses this new field to populate |specified_file_name| in mojom_files.mojom during serialization.

BUG=#609
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1577993003 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: dda8ed1cca593d935ec6910d4216ba716cbb75cc