Generate Mojom Types in Dart (Take 2)
This CL modifies the Dart generator to produce mojom type information.

This includes a test to ensure that the output is expected.

The previous CL (located at https://codereview.chromium.org/1433183002/)
has been replaced with this one.

In order for this to compile, hacks were added to the Dart embedder
package builder. This is acceptable because those files are supposed
to be removed in the future.

All .mojom.dart files were regenerated in this CL.

BUG=https://github.com/domokit/mojo/issues/596
R=azani@chromium.org, zra@google.com

Review URL: https://codereview.chromium.org/1539673003 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: a49d2b24078041ffa78e7e8d9c72ffed213f7881