Mojom parser: Introduce meta-data-only mode.
When the flag -meta-data-only is passed to the parser binary it will be in meta-data-only mode. This means the parser will:
(1) Only parse the file-level attributes and 'module' declaration and not any of the mojom declarations.
(2) The resolution and validation steps will be skipped.
(3) Imported file names will not be resolved to full paths and the imported files will not be parsed.

This is part of my work to delete the old parser. Currently mojo/dart/tools/presubmit/check_mojom_dart.py is invoking the old parser.py script directly in order to read the 'DartPackage' attributes and 'module' declarations of .mojom files. The old parser.py script, in conjunction with translate.py, was able to parse a single .mojom file without resolving and verifying it. So we need to provide similar functionality in order to delete the old script.

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1689013003 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 72a4569fa1c2319ef80d251e60971a59034d1551