Mojom compiler: Move RuntimeTypeInfo from mojom_files.mojom to mojom_types.mojom.
Previously it was awkward/impossible for mojom_types.mojom.go to include the functions that implement runtime type discovery because the structs were defined in mojom_files.mojom.go which is in a different package that depends on mojom_types.mojom.go. This was an unfortunate situation because it meant that any other mojom that depended on mojom_types.mojom also could not fully implement runtime type discovery.

BUG=#448
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1717583003 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 114885d95b658a7ef6d3225aaccd1be2b748b03c