Updated LiteralValue to hold a pointer to the token that represents the literal value.
This addresses a bug in the current printer. Namely, the current printer simply
does not know how a literal constant was formatted. This is problematic for
integers which could be specified in hexadecimal and floating point numbers
which could be specified in exponential notation. By storing the token, we know
exactly how the literal value was represented in the original source code.

The bulk of this CL is actually updating tests.

R=rudominer@chromium.org, rudominer
BUG= #592

Review URL: https://codereview.chromium.org/1721163002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 1851719b1c301df9d7236d1b12f947424a84e875