Mojom runtime type info: New implementation for Dart.
This is a follow on to https://codereview.chromium.org/1719873003/
and does for the Dart bindings the same thing that that change did for the Go bindings.

Our implementation of runtime type information is changing and this CL implements the change for Dart.

In the older implementation the runtime type structures are computed and constructed by the code generators (thus in Python and the Jinja 2 templates).
In this newer implementation the Mojom parser computes the information and stores it in a mojom_types.RuntimeTypeInfo
structure. Then the code generators insert a literal array of bytes representing a serialized RuntimeTypeInfo into the generated Dart code. At runtime all a Mojo application needs to do
is deserialize the RuntimeTypeInfo struct and use it to answer queries.

One particular change to note is that like with Go, the new implementation generates an implementation of getTopLevelInterface() that returns null for an interface that is not a top-level interface (i.e. is not marked with the "ServiceName" annotation.)

BUG=#448
R=alexfandrianto@google.com, azani@chromium.org, zra@google.com

Review URL: https://codereview.chromium.org/1753013002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 04e96bfd15364f988b745d76f17b59e214cb9cc6