Mojom runtime type info: Use compression and base64 encoding.
This change decreases the size of the data written into the generated source code that represents a serialized Mojo struct. Previously we were writing a literal array of bytes, written in decimal. Now we have done two things to decrease the size:
- We gzip the data
- We base64 encode the bytes and write an ascii string literal.
The result is that we decrease the size of the data written into the source by a factor of 30 in many cases.

BUG=#448
R=azani@chromium.org, zra@google.com

Review URL: https://codereview.chromium.org/1776243004 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 2d4da54579bbcc2b37c28393fab8e3b25aee5e7f