Go bindings: Initialize Mojom runtime type info lazily.
Rather than base64 decoding and decompressing and deserializing during init() we now do it lazily when the feature is first used. This makes it cheaper to include the code even if the feature is not being used. This is what we already do in Dart.

Since the deserialization doesn't occur in init() and there is no reason other than program corruption for the deserialization to fail, we now panic if there is a failure.

BUG=#448
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1780593004 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 921b78f86b55f88984fcc49b5cbef30b5d7c7fd6