Mojom parser: Stop populating the complete_typeset field.
The complete_typeset field is expensive to populate because it requires a depth-first search through the type graph rooted at each top-level interface. We originally thought this would be needed for runtime type information but it is currently not be used at all for that purpose. For now we are going to stop populating the field and add a comment that it is deprecated. We will likely remove it in the future.

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1786543002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: efa10ff9d0829d4a5df0eab34364d7e117b85022