Mojom compiler: Port pack tests to the frontend.
We add a new test, TestStructsComputedData, to computed_data_test.go, with a bunch of test cases that correspond to all of the test cases in pack_tests.py and pack_unittests.py. In a later CL we plan to delete those two python tests because we will be deleting the code in pack.py that they test because that logic is now being performed by the frontend instead.

I turns out that pack_tests.py is not actually being run and contains a bug. In this patch I also fix that bug. But I do not bother fixing the problem that it is not being run becuase, as I mentioned, I am about to delete it.

Finally I have eliminated a little duplicate code in computed_data.go.

BUG=#713
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1833593002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: b013d061e7904f5e8a26fe94e64481b01444886d