mojom_types.mojom: Add |min_version| fields in a few more places.
We add a |min_version| field to the following:
- UnionField
- EnumValue
- MojomMethod

We also add a |current_version| field to MojomInterface.

Some details about the above:
- Currently the code generators do not support the MinVersion feature for union fields or enum values.
However there was some stub of support for this that I deleted from module.py in https://codereview.chromium.org/1824263002/ so it appears that the MinVersion feature is intended to eventually be supported for union fields and enum values.

- The lack of a |min_version| field in MojomMethod was an oversight. This is needed and currently the old backend is parsing it from the "MinVersion" attribute.

- The |current_version| field for MojomInterface is something that is currently being computed in the backend and so adding this field is part of the effort to move more computations to the frontend.

Note that these new fields are not currently being populated or consumed. That will happen in a follow-up cl.

We also throw in some unrelated cleanups to mojom_types.mojom:

- Delete the unused |resolved_concrete_value| field from UserValueReference.
- Delete the unused |min_version| field from DeclarationData.
- Fix comments in several places.

BUG=#713

Review URL: https://codereview.chromium.org/1829403004 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: a33fe682700c1272dd908721c044c89f19905f91