Mojom compiler backend: Stop re-computing interface version numbers and consume method declaration order.
The current version of an interface is now being computed in the frontend so in this patch we stop re-computing this value in the backend and simply consume the version from the intermediate representation.

Similarly we stop parsing the "MinVersion" attribute for a method and consume the parsed value from the intermediate representation.

Finally we sort the methods of an interface by declaration order instead of by ordinal order.
A method's ordinal is specified in the |ordinal| field so with this change the generators will be able to know both the ordinal and the declaration order.

BUG=#713
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1847683004 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: f444558244bac885e7c9644608c010b25a9bfddd