Dart: Fix bug in unbind() for Stubs.
The MojoHandleWatcher.remove() in MojoEventSubscription's
unsubscribe() call can race with the removal of a handle
from the handle watcher for an arriving event.

This CL adds a test that recreates this situation, and
removes an assert that fires when it happens.

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1874553002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: aaeb3451ba80ead4583ee24773562aee4ef83ef5