Fix path logic for finding the jinja2 python module
If the mojom bindings generator is unable to find the jinja2 python
module, it attempts to add the bundled jinja2 module to the path so that
the import will not fail.

However, it attempts to find the path to the module by looking for the
directory "above" the directory named "public".  "public" is the name of
the mojo_sdk in the mojo tree but it could be named literally anything
if someone were to check out the mojo_sdk repo instead of the mojo repo.

Fix the path logic by assuming that the template_expander is in the
tools/bindings/pylib/mojom/generate directory and use a relative path to
find the jinja2 module.

R=vardhan@google.com, viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1886973007 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 6f2fc3aefc123696eddbd44e1b77f4477e15bca5