Mojom frontend: Start populating |resolved_concrete_value| field in mojom_types.mojom
In struct DeclaredConstant in mojom_types.mojom there is a field |resolved_concrete_value| that has been commented out and thus (obviously) not populated. The corresponding value has always been computed in the frontend Go code but we were not populating it in the intermediate representation because we were not aware of any reason to do so.

But now Vardhan will be able to use this in the new C generator. (Also I think some of the existing generators such as Dart will be able to stop computing this value themselves).

- I also added some additional tests of the existing computation of this value.

- I also fixed the name of an unrelated enum value in user_defined_types.go that I happened to notice was misnamed.

R=azani@chromium.org, vardhan@google.com

Review URL: https://codereview.chromium.org/1933563002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: f38d5d7e77ceb6b015d87d2d42bc4606090803df