Rust: Add initial bindings and examples
The first draft of Rust bindings for Mojo. The bindings themselves live
in a Cargo package so that in the future we may utilize Cargo, but
everything may be built with GN. Any and all feedback would be much appreciated!

BUG=
R=eholk@chromium.org, viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/2037333002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 382f5232d4edb033551c5a1fc785ac7393747265