C bindings: Implement *DecodePointersAndHandles() with some preliminary unittests.
This is basically an inverse of EncodePointersAndHandles() -- the code
for it is /very/ simiarly, with only the inverse operations (of
extracting handles and translating offsets->pointers).

Unittests for it involves saving the underlying buffer before
encoding, and then encoding+decoding, and checking that the underlying
buffer is the same as the one we saved.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/2126153002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: fedc14cacf280eac9ee5da692950a72d0b57f284