Rust: Add event loop
Implements a thread-local event loop similar to the RunLoop found in the C++ bindings. Currently, it does not support delayed tasks. The interface to the event loop is in writing handlers, structs, or types that implement the Handler trait. The Rust RunLoop lives in TLS but may also be created as a standalone object.

This event loop does not support nested event loop invocation. Currently however, one may create and invoke a *new* event loop from anywhere, which to an extent solves the nested event looping problem. The only restriction is that the user is limited to running the nested event loop on a disjoint sets of handlers from the parent event loop.

R=eholk@chromium.org

Review URL: https://codereview.chromium.org/2158723003 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 28404b7472508cc4b01a543dd9b9b6139a5160f8