Rust: Support delayed tasks in RunLoop
Adds the ability to schedule a delayed task in the RunLoop.
The task is inserted as a closure and as long as one handle
lives in the RunLoop it is guaranteed to run at the confirmed
deadline (with no handlers the RunLoop exits).

R=eholk@chromium.org

Review URL: https://codereview.chromium.org/2244463002 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: fefa014de0c9b55547ea19ed005b252a00dbe2e2