[dart] Update mojom template and public interfaces for new Dart scheme
This updates the mojom template to generate a dart package and fixes
up the DartPackage annotations to match the new package naming scheme.
Dart consumers of a target called mojom("foo") in //a/b/BUILD.gn must
declare a dependency on the generated //a/b/foo:foo_dart target.

The DartPackage annotations are now redundant as the dart package is
specified in the build file, but to simplify the transition I've
simply updated the annotation in the most broadly used mojom files and
not attempted to patch the generator. I'll look at that soon.

Change-Id: I4df20018714e30b3882436cf33740ac5ef82bb22